+++ src/test/java/com/jakubsuszynski/apiclient/RestClientTest.java
@@ -4,6 +4,8 @@ import com.jakubsuszynski.restresponse.RestResponse;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
@@ -22,7 +24,7 @@ public class RestClientTest {
     @Test
     public void checkCurrencyRates() {
         restResponse = restClient.getCurrencyRates("eur 2017-11-20 2017-11-24");
-        assertThat(restResponse.getRates().get(3).getBid(), is(4.1721));
+        assertThat(restResponse.getRates().get(3).getBid(), is(BigDecimal.valueOf(4.1721)));
     }
 
     @Test
